consolidate converter into system
