--- frontend/src/lib/services/api/converter-api-service.ts
@@ -1,10 +1,0 @@
-import BaseAPIService from './api-service';
-
-export default class ConverterAPIService extends BaseAPIService {
-	async convert(dockerRunCommand: string) {
-		const res = await this.api.post('/convert', {
-			dockerRunCommand
-		});
-		return res.data;
-	}
-}
