+++ frontend/src/routes/+layout.ts
@@ -1,4 +1,3 @@
-import { environmentManagementAPI } from '$lib/services/api';
 import { environmentStore } from '$lib/stores/environment.store';
 import versionService from '$lib/services/version-service';
 import { tryCatch } from '$lib/utils/try-catch';
@@ -8,6 +7,7 @@ import type { SearchPaginationSortRequest } from '$lib/types/pagination.type';
 import type { AppVersionInformation } from '$lib/types/application-configuration';
 import { userService } from '$lib/services/user-service';
 import { settingsService } from '$lib/services/settings-service';
+import { environmentManagementService } from '$lib/services/env-mgmt-service';
 
 export const ssr = false;
 
@@ -27,7 +27,7 @@ export const load = async () => {
 
 	const environmentsPromise = userPromise.then(async (user) => {
 		if (!environmentStore.isInitialized() && user) {
-			const environments = await tryCatch(environmentManagementAPI.getEnvironments(environmentRequestOptions));
+			const environments = await tryCatch(environmentManagementService.getEnvironments(environmentRequestOptions));
 			if (!environments.error) {
 				await environmentStore.initialize(environments.data.data, true);
 			}
