+++ frontend/src/routes/environments/[id]/+page.ts
@@ -1,8 +1,8 @@
+import { environmentManagementService } from '$lib/services/env-mgmt-service';
 import type { PageLoad } from './$types';
-import { environmentManagementAPI } from '$lib/services/api';
 
 export const load: PageLoad = async ({ params }) => {
-	const environment = await environmentManagementAPI.get(params.id);
+	const environment = await environmentManagementService.get(params.id);
 
 	return {
 		environment
